requirements.txt
# ===== NEXUSAI REQUIREMENTS =====
# Consolidated requirements for NexusAI application
# Supports both minimal and full ML functionality

# ===== CORE FRAMEWORK =====
Flask>=2.3.0,<4.0.0
Werkzeug>=2.3.0,<4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask-cors>=4.0.0

# ===== AI/ML CORE =====
groq>=0.33.0

# ===== MULTI-PROVIDER LLM SUPPORT (Optional) =====
# Uncomment the providers you want to use
# openai>=1.0.0                    # For GPT models
# anthropic>=0.25.0                # For Claude models  
# google-generativeai>=0.3.0       # For Gemini models
# requests>=2.31.0                 # For Ollama and Hugging Face (already included above)

# ===== DATABASE (Optional - for production) =====
# Uncomment these for production database support
# SQLAlchemy>=2.0.0
# Flask-SQLAlchemy>=3.1.0
# Flask-Migrate>=4.0.0
# psycopg2-binary>=2.9.0
# alembic>=1.12.0

# ===== CACHING & PERFORMANCE (Optional) =====
# Uncomment for production caching
# Flask-Caching>=2.1.0
# redis>=5.0.0

# ===== SECURITY & PRODUCTION (Optional) =====
# Uncomment for production security features
# Flask-Limiter>=3.5.0
# cryptography>=41.0.0
# PyJWT>=2.8.0

# ===== PRODUCTION SERVER (Optional) =====
# Uncomment for production deployment
# gunicorn>=21.2.0
# gevent>=23.9.0
# uwsgi>=2.0.0

# ===== AI/ML ADVANCED FEATURES (Optional) =====
# Uncomment these for full RAG/LoRA functionality
# torch>=2.6.0
# transformers>=4.40.0
# sentence-transformers>=2.7.0
# faiss-cpu>=1.8.0
# peft>=0.11.0
# accelerate>=0.30.0

# ===== LANGCHAIN ECOSYSTEM (Optional) =====
# Uncomment for advanced RAG features
# langchain>=0.2.0
# langchain-community>=0.2.0
# langchain-core>=0.2.0

# ===== VECTOR DATABASES (Optional) =====
# Uncomment for vector search capabilities
# chromadb>=0.5.0
# numpy>=1.24.0
# scikit-learn>=1.3.0

# ===== DOCUMENT PROCESSING (Optional) =====
# Uncomment for document upload/processing
# pypdf2>=3.0.0
# python-docx>=1.1.0
# beautifulsoup4>=4.12.0
# markdown>=3.5.0

# ===== ML UTILITIES (Optional) =====
# Uncomment for advanced ML features
# datasets>=2.18.0
# tokenizers>=0.19.0
# huggingface-hub>=0.23.0

# ===== OPTIMIZATION & COMPRESSION (Optional) =====
# Uncomment for production optimization
# Flask-Compress>=1.14
# Pillow>=10.1.0

# ===== MONITORING & LOGGING (Optional) =====
# Uncomment for production monitoring
# sentry-sdk[flask]>=1.38.0
# prometheus-flask-exporter>=0.23.0
# structlog>=23.2.0

# ===== UTILITIES (Optional) =====
# Uncomment for additional utilities
# python-dateutil>=2.8.0
# pytz>=2023.3
# marshmallow>=3.20.0
# click>=8.1.0

# ===== DEVELOPMENT TOOLS (Optional) =====
# Uncomment for development/testing
# pytest>=7.4.0
# pytest-flask>=1.3.0
# pytest-cov>=4.1.0
# black>=23.11.0
# flake8>=6.1.0
# mypy>=1.7.0

# ===== INSTALLATION PROFILES =====
# 
# MINIMAL INSTALLATION (Basic functionality):
# pip install Flask Werkzeug python-dotenv requests flask-cors groq
#
# FULL INSTALLATION (All features):
# Uncomment all optional sections above and run: pip install -r requirements.txt
#
# PRODUCTION INSTALLATION:
# Uncomment Core, Database, Security, Production Server sections